concatBytes moved to Helpers
